requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # For PostgreSQL

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# GeoIP
geoip2==4.7.0
maxminddb==2.4.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
pandas==2.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3